fix(auth): split Signature from connection
